Added support for paths with spaces, and bats tests.
The maven wrapper script currently does not work in folder heirarchies that contain spaces.  The simple fix was to quote the $(pwd) statements in find_maven_basedir along with some more quote-fu on the MAVEN_PROJECTBASEDIR and MAVEN_OPTS line, but I opted for a change that would be testable instead.

This involved parameterizing find_maven_basedir, and reorganizing the script to make the actual body of it excludable by bats when it is loaded.

My apologies as this will make the PR a bit messier than need be for this change, but I will do my best to highlight the most relevant bits.

I did incorporate the bats run as part of the automation launched when building yet as that would require a dependency on an external toolset and should be probably be discussed first.