Use project-specific maven settings.xml
When connecting through a corporate proxy, the maven settings must include proxy details.  
There was an existing implementation to set the java system properties using jvm.config.  But, this didn't set _maven_'s proxy configuration.  

In this change, the mvnw batch file checks for the existence of a file called `settings.xml` in the `%MAVEN_PROJECTBASEDIR%\.mvn` directory.  If a file is found there, the --settings parameter is passed at the end of the maven command line.  
If no such file is found, the parameter is not included.