Get maven user home (~/.m2) with the same logic with maven
Now wrapper will guess the maven user home location via system props, system envs
and at last the default location (`~/.m2`). This order is the convention of the gradle.
For maven, user home has a fixed location, i.e. `~/.m2`.

For wrapper, we can still use `zipBase` and `distributionBase` to directly custom
the location of the downloaded and installed artifacts.