mvnw.cmd: use 'usebackq' for parsing files
When parsing the properties file, the default for-loop with the form `for /F ... ("path")` will
treat "path" as a literal string, not a file. On the other hand, the `path` may contain spaces.
After switching to the form `for /F "usebackq ..." ... ("path")`, `mvnw.cmd` can parse the
config file at "path" correctly.